Separate web and io implementations of network image (#34112)
* add web and io implemenations of network and asset image

* fix foundation import

* update to remove extra asset image indirection

* skip chunk test

* address comments

* disable non-functional test

* disable all golden tests

* address comments
