+++ packages/flutter/lib/src/painting/_network_image_web.dart
@@ -1,0 +1,73 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'dart:async';
+import 'dart:ui' as ui;
+
+import 'package:flutter/foundation.dart';
+
+import 'image_provider.dart' as image_provider;
+import 'image_stream.dart';
+
+/// The dart:html implemenation of [image_provider.NetworkImage].
+class NetworkImage extends image_provider.ImageProvider<image_provider.NetworkImage> implements image_provider.NetworkImage {
+  /// Creates an object that fetches the image at the given URL.
+  ///
+  /// The arguments [url] and [scale] must not be null.
+  const NetworkImage(this.url, {this.scale = 1.0, this.headers})
+      : assert(url != null),
+        assert(scale != null);
+
+  @override
+  final String url;
+
+  @override
+  final double scale;
+
+  @override
+  final Map<String, String> headers;
+
+  @override
+  Future<NetworkImage> obtainKey(image_provider.ImageConfiguration configuration) {
+    return SynchronousFuture<NetworkImage>(this);
+  }
+
+  @override
+  ImageStreamCompleter load(image_provider.NetworkImage key) {
+    return MultiFrameImageStreamCompleter(
+      codec: _loadAsync(key),
+      scale: key.scale,
+      informationCollector: () {
+        return <DiagnosticsNode>[
+          DiagnosticsProperty<image_provider.ImageProvider>('Image provider', this),
+          DiagnosticsProperty<NetworkImage>('Image key', key),
+        ];
+      },
+    );
+  }
+
+  Future<ui.Codec> _loadAsync(NetworkImage key) async {
+    assert(key == this);
+
+    final Uri resolved = Uri.base.resolve(key.url);
+    // This API only exists in the web engine implementation and is not
+    // contained in the analyzer summary for Flutter.
+    return ui.webOnlyInstantiateImageCodecFromUrl(resolved); // ignore: undefined_function
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    if (other.runtimeType != runtimeType) {
+      return false;
+    }
+    final NetworkImage typedOther = other;
+    return url == typedOther.url && scale == typedOther.scale;
+  }
+
+  @override
+  int get hashCode => ui.hashValues(url, scale);
+
+  @override
+  String toString() => '$runtimeType("$url", scale: $scale)';
+}
