+++ packages/flutter/lib/src/painting/image_resolution.dart
@@ -5,7 +5,6 @@
 import 'dart:async';
 import 'dart:collection';
 import 'dart:convert';
-import 'dart:io';
 import 'dart:ui' show hashValues;
 
 import 'package:flutter/foundation.dart';
@@ -265,10 +264,13 @@ class AssetImage extends AssetBundleImageProvider {
       return _naturalResolution;
     }
 
-    final File assetPath = File(key);
-    final Directory assetDir = assetPath.parent;
+    final Uri assetUri = Uri.parse(key);
+    String directoryPath = '';
+    if (assetUri.pathSegments.length > 1) {
+      directoryPath = assetUri.pathSegments[assetUri.pathSegments.length - 2];
+    }
 
-    final Match match = _extractRatioRegExp.firstMatch(assetDir.path);
+    final Match match = _extractRatioRegExp.firstMatch(directoryPath);
     if (match != null && match.groupCount > 0)
       return double.parse(match.group(1));
     return _naturalResolution; // i.e. default to 1.0x
